third_party/blink/web_tests/fast/forms/fieldset/legend-absolute-position-auto-width-expected.txt

This test checks the absolute postioned legend element to honor auto width.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS legend.offsetWidth is 200
PASS legend.clientWidth is 200
PASS successfullyParsed is true

TEST COMPLETE

Legend
Fieldset content
